** exercise003.js ** Code implementation for 'getTotalSubjects'.
Note
Found a nice code snippet to return a unique list of subjects, routine was not required for the test case to pass but the wording of the test description implies that the code should return the number of UNIQUE subjects.  Music is taken by 2 students, but music is a single subject and should not be counted twice.